refactored pds url resolution into its own file, and made use of it for video downloads
